After long and grueling work all is finished
No more commit messages as it was done under the pressure of time and
the code is less than pretty in some places. Hopefully I come back
and refactor. All seems to work just fine, although some normals in
city *.obj file used by this app seem to be wrong.